requirements.txt
python-dotenv>=1.0.0
clarifai>=9.10.0
